# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: LyceeProject API
  description: |
    API pour la plateforme de mise en relation entre lycées professionnels et entreprises.
    
    Cette API permet aux entreprises de créer des demandes de partenariat et aux lycées 
    de consulter et répondre à ces demandes. Elle inclut également des fonctionnalités 
    de matching automatique basées sur des critères géographiques et sectoriels.
    
    ## Authentification
    L'API utilise l'authentification JWT (JSON Web Token). Après connexion, incluez le token 
    dans l'en-tête Authorization: `Bearer <token>`
    
    ## Types d'utilisateurs
    - **Entreprise** : Peut créer des demandes de partenariat
    - **RBDE** (Lycée) : Peut consulter et traiter les demandes
    
  version: 1.0.0
  contact:
    name: Support LyceeProject
    email: support@lyceeproject.fr
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
  - url: https://api.lyceeproject.fr/api
    description: Serveur de production
security:
  - BearerAuth: []
paths:
...
  /demandes:
    get:
      tags:
        - Demandes
      summary: Liste des demandes
      description: Récupère la liste des demandes de partenariat avec filtres
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Numéro de page
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Nombre d'éléments par page
        - in: query
          name: statut
          schema:
            type: string
            enum: [EN_ATTENTE, EN_COURS, TRAITE, ANNULE]
          description: Filtrer par statut
        - in: query
          name: priorite
          schema:
            type: string
            enum: [BASSE, NORMALE, HAUTE, URGENTE]
          description: Filtrer par priorité
        - in: query
          name: entreprise_id
          schema:
            type: string
          description: Filtrer par entreprise
        - in: query
          name: search
          schema:
            type: string
          description: Recherche textuelle
      responses:
        '200':
          description: Liste des demandes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      demandes:
                        type: array
                        items:
                          $ref: '#/components/schemas/Demande'
                      total:
                        type: integer
                        example: 50
                      page:
                        type: integer
                        example: 1
                      totalPages:
                        type: integer
                        example: 5

    post:
      tags:
        - Demandes
      summary: Créer une demande
      description: Créer une nouvelle demande de partenariat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entreprise_id:
                  type: string
                  example: "123456789"
                  description: "ID ou SIRET de l'entreprise"
                titre:
                  type: string
                  example: "Stage en développement web"
                description:
                  type: string
                  example: "Recherche de stagiaires en développement web"
                type_partenariat:
                  type: string
                  example: "STAGE"
                  enum: [STAGE, APPRENTISSAGE, PROJET, INTERVENTION]
                zone_geo:
                  type: string
                  example: "Île-de-France"
                nb_places:
                  type: integer
                  example: 5
                date_debut_souhaitee:
                  type: string
                  format: date
                  example: "2024-03-01"
                date_fin_souhaitee:
                  type: string
                  format: date
                  example: "2024-06-30"
                priorite:
                  type: string
                  enum: [BASSE, NORMALE, HAUTE, URGENTE]
                  default: NORMALE
                metier_id:
                  type: string
                  example: "uuid-metier"
              required:
                - entreprise_id
                - titre
      responses:
        '201':
          description: Demande créée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Demande créée avec succès"
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "uuid-demande"
  /demandes/{id}:
    get:
      tags:
        - Demandes
      summary: Détail d'une demande
      description: Récupère les détails d'une demande spécifique
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID de la demande
      responses:
        '200':
          description: Détails de la demande
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DemandeDetail'

    put:
      tags:
        - Demandes
      summary: Modifier une demande
      description: Modifie une demande existante
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                titre:
                  type: string
                description:
                  type: string
                statut:
                  type: string
                  enum: [EN_ATTENTE, EN_COURS, TRAITE, ANNULE]
                priorite:
                  type: string
                  enum: [BASSE, NORMALE, HAUTE, URGENTE]
      responses:
        '200':
          description: Demande modifiée avec succès
  /demandes/{id}/lycees:
    post:
      tags:
        - Demandes
      summary: Assigner des lycées
      description: Assigne des lycées à une demande
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lycee_ids:
                  type: array
                  items:
                    type: string
                  example: ["uuid-lycee-1", "uuid-lycee-2"]
                note:
                  type: string
                  example: "Lycées sélectionnés pour leur expertise"
              required:
                - lycee_ids
      responses:
        '200':
          description: Lycées assignés avec succès
  /demandes/{id}/actions:
    get:
      tags:
        - Demandes
      summary: Historique des actions
      description: Récupère l'historique des actions sur une demande
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Historique des actions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Action'

  # ================================
  # LYCÉES
  # ================================
  /lycees:
    get:
      tags:
        - Lycées
      summary: Liste des lycées
      description: Récupère la liste des lycées avec filtres
      parameters:
        - in: query
          name: region
          schema:
            type: string
          description: Filtrer par région
        - in: query
          name: departement
          schema:
            type: string
          description: Filtrer par département
        - in: query
          name: search
          schema:
            type: string
          description: Recherche par nom
      responses:
        '200':
          description: Liste des lycées
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lycee'
...
  /lycees/{uai}:
    get:
      tags:
        - Lycées
      summary: Détail d'un lycée
      description: Récupère les détails d'un lycée par son code UAI
      parameters:
        - in: path
          name: uai
          required: true
          schema:
            type: string
          description: Code UAI du lycée
          example: "0751234A"
      responses:
        '200':
          description: Détails du lycée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/LyceeProfile'

  # ================================
  # ENTREPRISES
  # ================================
...
  /stats/demandes:
    get:
      tags:
        - Statistiques
      summary: Statistiques des demandes
      description: Récupère les statistiques sur les demandes de partenariat
      parameters:
        - in: query
          name: entreprise_id
          schema:
            type: string
          description: Filtrer par entreprise
        - in: query
          name: date_debut
          schema:
            type: string
            format: date
          description: Date de début pour le filtre
        - in: query
          name: date_fin
          schema:
            type: string
            format: date
          description: Date de fin pour le filtre
      responses:
        '200':
          description: Statistiques des demandes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StatsDemandes'

  # ================================
  # DONNÉES DE RÉFÉRENCE
  # ================================
...
  /sync/lycees:
    post:
      tags:
        - Synchronisation
      summary: Synchroniser les lycées
      description: Synchronise les données des lycées depuis les sources externes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
                  description: "Forcer la synchronisation même si les données sont récentes"
      responses:
        '200':
          description: Synchronisation réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      lycees_synchronises:
                        type: integer
                      nouveaux_lycees:
                        type: integer
  /sync/entreprise/{siret}:
    post:
      tags:
        - Synchronisation
      summary: Synchroniser une entreprise
      description: Synchronise les données d'une entreprise depuis l'API Sirene
      parameters:
        - in: path
          name: siret
          required: true
          schema:
            type: string
          description: Numéro SIRET de l'entreprise
      responses:
        '200':
          description: Synchronisation réussie

  # ================================
  # UTILITAIRES
  # ================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Demande:
      type: object
      properties:
        id:
          type: string
          example: "uuid-demande"
        titre:
          type: string
          example: "Stage en développement web"
        description:
          type: string
          example: "Recherche de stagiaires en développement web"
        type_partenariat:
          type: string
          enum: [STAGE, APPRENTISSAGE, PROJET, INTERVENTION]
          example: "STAGE"
        statut:
          type: string
          enum: [EN_ATTENTE, EN_COURS, TRAITE, ANNULE]
          example: "EN_ATTENTE"
        priorite:
          type: string
          enum: [BASSE, NORMALE, HAUTE, URGENTE]
          example: "NORMALE"
        zone_geo:
          type: string
          example: "Île-de-France"
        nb_places:
          type: integer
          example: 5
        date_debut_souhaitee:
          type: string
          format: date
          example: "2024-03-01"
        date_fin_souhaitee:
          type: string
          format: date
          example: "2024-06-30"
        date_creation:
          type: string
          format: date-time
        date_modification:
          type: string
          format: date-time
        entreprise:
          $ref: '#/components/schemas/Entreprise'
        metier:
          $ref: '#/components/schemas/Metier'
    DemandeDetail:
      allOf:
        - $ref: '#/components/schemas/Demande'
        - type: object
          properties:
            lycees:
              type: array
              items:
                type: object
                properties:
                  lycee:
                    $ref: '#/components/schemas/Lycee'
                  statut_traitement:
                    type: string
                    enum: [NOUVEAU, EN_COURS, ACCEPTE, REFUSE]
                  score_matching:
                    type: number
                    format: float
                    example: 0.85
                  distance_km:
                    type: number
                    format: float
                    example: 15.5
                  note:
                    type: string
                  date_affectation:
                    type: string
                    format: date-time
            actions:
              type: array
              items:
                $ref: '#/components/schemas/Action'
    Lycee:
      type: object
      properties:
        id:
          type: string
          example: "uuid-lycee"
        nom:
          type: string
          example: "Lycée Professionnel Jean Dupont"
        numero_uai:
          type: string
          example: "0751234A"
        type_etablissement:
          type: string
          example: "Lycée professionnel"
        statut_public_prive:
          type: string
          enum: [Public, Privé]
          example: "Public"
        adresse:
          type: string
          example: "123 Rue de la République"
        code_postal:
          type: string
          example: "75001"
        commune:
          type: string
          example: "Paris"
        departement:
          type: string
          example: "Paris"
        region:
          $ref: '#/components/schemas/Region'
        latitude:
          type: number
          format: float
          example: 48.8566
        longitude:
          type: number
          format: float
          example: 2.3522
        telephone:
          type: string
          example: "01.23.45.67.89"
        email:
          type: string
          format: email
          example: "contact@lycee.ac-paris.fr"
        site_web:
          type: string
          format: uri
          example: "https://lycee.ac-paris.fr"
        formations:
          type: array
          items:
            $ref: '#/components/schemas/Formation'
...
    Action:
      type: object
      properties:
        id:
          type: string
          example: "uuid-action"
        type_action:
          type: string
          enum: [CREATION, MODIFICATION, AFFECTATION, STATUT_CHANGE]
          example: "CREATION"
        commentaire:
          type: string
          example: "Demande créée"
        date_action:
          type: string
          format: date-time
        user:
          type: object
          properties:
            full_name:
              type: string
              example: "Jean Dupont"
            email:
              type: string
              format: email
              example: "jean.dupont@example.com"
        donnees_avant:
          type: object
          description: "État avant modification (JSON)"
        donnees_apres:
          type: object
          description: "État après modification (JSON)"
...